Reviewer 2 comments -----
Fixed Pag. 2 l+7: E has not been introduced in a clear way.
   - Added missing subscript to E (this was done during 
     prior rounds of editing)

Fixed Pag. 4 l+5: The supra-index of the equation are ok?
   - Super-scripts corrected (moved from f to h for both sides
     of the equation) 

Fixed Section 2.1: Is convex the optimization problem?
   - Made it clear that this is *not* a convex optimization problem
     even though the surface defined by the classifier score is convex.

Skipped Section 4: A table with a summary of the characteristics of the
corpora should be introduced.
   - There's no way we can fit this in with rewritting sizable 
     amounts of the paper to make things shorter.
     (may fix this if there's time)
  

Fixed Table 3: I suppose that the figures reported in the table are bleu
scores.
   - Fixed for tables 3,4, & 5 - we now explicitly mention BLEU scores 
     for each.

Section 5, end of paragraph 2: I'm amazing that the differences are
statistical significant.
   - yeah, well they are, and very significant at that.


svn path=/trunk/javanlp/; revision=19834