minor misc. fix:
- Changed subscripts on table 2 to superscripts in order to be consistent with other tables

Also, I'd like to note that all of reviewer three's minor comments have been addressed:

Fixed - Och's paper on Minimum Error Rate Training appeared in 2003, not 2002.
   - one of the first things that was fixed

Fixed -The colors in Figs. 1 & 2 simply do not convey in my black & white printed
copy.  Many people will read the paper this way, so please make sure that the
figures are understandable in this medium.
    - again, one of the first things that was fixed

Fixed - I am not sure whether it is necessary to go into the implementation details of the source code, but in any case I think it would be fair to credit David
Chiang, the  author of the optimization code included with Moses.  The KCD
method is based on Philipp Koehn's older perl training script for the Pharoah
toolkit; it is basically the algorithm described in Och's 2003 paper.
      - fixed earlier. Footnote now includes mention of David Chiang


svn path=/trunk/javanlp/; revision=19835