Reviwer 2 big comments (1)
Fixed -The first paragraph states that Och's algorithm works by "walking the true
error surface".  This is not really true.  Och's algorithm samples a single
point in the parameter space and constructs an *approximation* to the true
error surface using n-best lists.  It is this approximate surface that is
walked.  However, as the the distance from the sampled point increases, the
correctness of this approximation dwindles, and thus the optimum on the surface
might be very far from the true optimum due to errors in the approximation
(this is why early iterations of the algorithm often return very poor results).
 Although the approximation improves over time as more points are sampled, it
is important to keep in mind that the optimization always occurs on this
approximate surface.

    - Removed offending language from the introduction
    - Section 2 now includes details regarding how MERT training is
      performed iteratively, using cumulative n-best lists to 
      approximate the hypothesis space.
      

-What is called here "regularization" is really a form of smoothing.  The use
of the term "regularization" is probably fine, as I understand it encompasses
methods like this.  However, in most of the recent literature at *CL
conferences it seems to refer specifically to methods used to control or
minimize the number of parameters, which is not done here.  I was initially a
bit confused by the usage.
   
    - Added footnote to address this


svn path=/trunk/javanlp/; revision=19837