last of reviewer comments addressed....
review 3

-The regularization method aims to solve the same problem as the annealing
method presented by Smith & Eisner (ACL-COLING 2006).  I wonder how the two
methods compare?  Even if an experimental comparison is not possible, some
discussion the similarity should be included in the paper.  The Smith & Eisner
paper should be cited.
  
   - Added cite, added discussion

-The algorithm implemented in the Moses toolkit follows Och (2003) by using
random starting points for the optimization (this an ad hoc method used for
much the same reasons as annealing or this paper's regularization method).  The
dirty little secret of this algorithm is that the randomization can result in
some quite different results for test set BLEU just by re-running it on the
exact same development set.  Unfortunately, the statistical significance test
doesn't address this.  Since this paper studies the optimization method
directly, it might be more convincing if the results in Tables 3, 4 & 5 were
generated using some kind of average over multiple runs.  Otherwise, there's no
way to know whether the results are simply the result of a bit of luck during
the generation of random points in the optimization process.
 
   - added some comments on this


svn path=/trunk/javanlp/; revision=19853