Turn all of the static state in GlobalHolder into member variables, which affects a ton of other classes in the tagger and using the tagger.  There is still some work to do to make the tagger re-entrant and to clean up the code, but this is most of the way there already.  One obvious change that needs doing is getting rid of the GlobalHolder class altogether and making it part of the MaxentTagger class, but I wanted to check this in and then make some separate changes to have a record of the steps in SVN.
Testing done: the new code builds left3words-distsim-wsj-0-18.tagger and left3words-wsj-0-18.tagger exactly the same as the old code, and it produces the exact same results when testing on a large data file using both of those taggers.

Testing not done: testing other projects that rely on the tagger.

svn path=/trunk/javanlp/; revision=29244